Refactor finance report serializers
