+++ app/models/reports/monthly_finance_report_serializer.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Reports
-  class MonthlyFinanceReportSerializer < BaseFinanceReportSerializer
+  class MonthlyFinanceReportSerializer < BaseSerializer
     ATTRIBUTES = {
       period: "period",
       year: "year",
@@ -44,5 +44,24 @@ module Reports
       chg_irimport_cnt: "chg_irimport_cnt",
       chg_irimport_tot: "chg_irimport_tot"
     }.freeze
+
+    def initialize(rows)
+      @rows = rows
+    end
+
+    def scope
+      @rows
+    end
+
+    def parse_object(results_row)
+      self.class::ATTRIBUTES.map do |key, _value|
+        presenter = FinanceReportPresenter.new(results_row)
+        if presenter.respond_to?(key)
+          presenter.public_send(key)
+        else
+          presenter[key]
+        end
+      end
+    end
   end
 end
